[FLINK-29729][parquet] Fix credential info configured in flink-conf.yaml is lost during creating ParquetReader
This closes #21966
